refactor getOne and deleteOne as handleOne

+++ src/utils/handler/refactor.handler.js
@@ -1,35 +1,34 @@
 import { AppError } from "../error/appError.js";
 import { catchAsyncError } from "../error/asyncError.js";
 
-const getOne = (model, docName) => {
+const handleOne = (model) => {
   return catchAsyncError(async (req, res, next) => {
     const { id } = req.params;
-    const existingDoc = await model.findById(id);
+    let doc;
 
-    if (!existingDoc)
-      return next(new AppError(`${docName} with ID ${id} not found`, 404));
+    if (req.method === "GET") {
+      doc = await model.findById(id);
+    } else if (req.method === "DELETE") {
+      doc = await model.findByIdAndDelete(id);
+    }
 
-    res.status(200).json({
-      status: "success",
-      data: existingDoc,
-    });
+    if (!doc)
+      return next(
+        new AppError(`${model.modelName} with ID ${id} not found`, 404)
+      );
+
+    if (req.method === "GET") {
+      return res.status(200).json({
+        status: "success",
+        data: doc,
+      });
+    } else if (req.method === "DELETE") {
+      return res.status(200).json({
+        status: "success",
+        message: `${doc.name} successfully deleted`,
+      });
+    }
   });
 };
 
-const deleteOne = (model, docName) => {
-  return catchAsyncError(async (req, res, next) => {
-    const { id } = req.params;
-
-    const deletedDoc = await model.findByIdAndDelete(id);
-
-    if (!deletedDoc)
-      return next(new AppError(`${docName} with ID ${id} not found`, 404));
-
-    res.status(200).json({
-      status: "success",
-      message: `${deletedDoc.name} successfully deleted`,
-    });
-  });
-};
-
-export { getOne, deleteOne };
+export { handleOne };
